requirements.txt
typing-extensions
sniffio
python-dotenv
idna
greenlet
annotated-types
sqlalchemy
anyio
starlette
fastapi
passlib
bcrypt
psycopg
psycopg2
uvicorn